22 January, 2001 - Jay Larson <larson@mcs.anl.gov>
Modifications for the port to the SunOS platform.  The changes made
were:

	1) addition of Makefile.conf.SunOS to the mpeu and mpeu/mpi0
	directories;
	2) minor corrections to the module m_zeit.F90, which were all
	replacement of 'x' format descriptors with '1x', which is
	apparently the f90 standard (though the first compiler we've
	seen where this is detected).

N.B.:  The make configuration used is mpich for MPI, with the mpich
path names correct on MCS machines fire and pitcairn.

Another Note:  Usually one tries to use the -fast option when
compiling f90 and c code.  I couldn't do this with f90, because
in some modules the compiler would pick -dalign, but in a higher-
level module (on that "uses" the aforementioned module), it would
not, and this generates an error.  For this reason, I've gone only
with the -O2 optimization flag.  In the future, somebody may wish
to investigate a better set of optimization flags.