22 January 2001 Jay Larson <larson@mcs.anl.gov>
Port of mct to the SunOS platform.  The main changes were the addition
of a Makefile.conf.SunOS, with -M flags included (this is how sun's
f90 compiler locates modules).  A minor change to m_Router.F90 was
required to clear up for the f90 compiler how to process the interface
blocks.

As with mpeu, this port was done using the local mpich installation,
which is good for both pitcairn and fire at MCS.  See the check-in
log for mpeu for more details...