07 February 2001 Jay Larson <larson@mcs.anl.gov>
Numerous changes marked this check-in:

1) The subroutine accumulate() is now completely general, which
required some changes to both the Accumulator, and the addition
of support for cross-indexing AttrVect types, Accumulators, and
Accumulator AttrVect combinations.

2) Implementation in m_ExchangeMaps of a prototype method that
allows two components to exchange GlobalSegMaps (resulting in
each having the other's GlobalSegMap).

3) Addition of support for cross-indexing of AttrVect lists,
Accumulator attribute lists, and general cross-indexing of the
elements present in the List datatype.  These are all implemented
in the new module m_SharedAttrIndices.