25 February, 2001 - J. Larson <larson@mcs.anl.gov>
Numerous changes and additions:

Changes in Makefile.conf.Linux-alpha to use mpif90 rather than f90.

Addition of a prototype AttrVect gather based on the GlobalSegMap,
and a bug fix the the AttrVect scatter (m_AttrVectComms.F90).

Added services to the GlobalSegMap, including:  parallel
initialization methods (2) that use _replicated_ decomposition
description data; a method to compute the number of points owned
by a particular process of a GlobalSegMap (ProcessStorage()).

A new module for converting the GlobalMap to a GlobalSegMap (always
permissible) and vice-versa (not always permissible--this method is
for now an unfinished prototype).