April 26, 2001 - J.W. Larson <larson@mcs.anl.gov>
A major logic bug fix in the routines ByRowGSMap_() and ByColumnGSMap_()
in the module m_SparseMatrixDecomp.F90.  This bug fix resulted in the
addition of two arguments to the interfaces of these routines.

This change is reflected to the changes in the calls to the above routines
in the module m_SparseMatrixComms.