13 June 2001 - Jay Larson
Atonement for sins against consistency.

This Check-in represents changes to make consistent the handling of the
status/stat output flag in the MCT derived type comms routines.  Now, all
the status/stat flags are OPTIONAL arguments.  Furthermore, if the argument
stat/status is present, it is now initialized to zero (success).  Sadly,
this was previously not the case.  Furthermore, changes have been made
to make consistent error handling in these routines.  If an exception
occurs that results in a non-zero value of stat/status, a message is
written to stderr, and the routine returns with stat/status set to the
appropriate nonzero value.