14 June 2001 - J. Larson and E. Ong
Repaired a bug in AttrVect_Sort() found by Everest Ong.  This routine
was sorting the INTEGER attributes correctly, but not the REAL
attributes.

While we had this routine cracked open, we took the opportunity to
clean up the error handling / shutdown logic to cover correctly all
the combinations of arguments perrWith/dieWith that might be used upon
invocation.