10 October, 2001 - J. Larson <larson@mcs.anl.gov>
Added an optional argument to the data-local SparseMatrix-AttrVect
multiply to allow the user to choose whether or not to multiply the
INTEGER attributes of the AttrVect by the SparseMatrix.  This serves
two purposes:  1) it preserves indexing information we use for
debugging purposes, and 2) it saves us the cost of cross-indexing for
the integer part of the multiply if we don't want to perform it
anyway.