26 September, 2002 - J. Larson
This check-in marks absorption into MCT of previously-tested
strategies for distributed SparseMatrix - AttrVect multiplication.
What had been a series of operations in the unit testers are now
encapsulated in the SparseMatrixPlus datatype, which is the master
class for distributed SparseMatrix - AttrVect multiplication.  This
Class and support functions are found in the new module
m_SparseMatrixPlus.

The addition of the SparseMatrixPlus solves the open problem of how
we were going to support distributed matrix-vector multiplication
in MCT.  There is now one new routine in m_MatAttrVectMul that uses
the SparseMatrixPlus, and the numerous, speculative, empty APIs in
this module have been excised.

Since a new module was added, it had to be listed in SRCS_F90.mk.

Also, a corresponding definition of its dependencies had to be
added to Makefile.dependency.  This process also unveiled a missing
dependency that was also corrected.