22 July, 2003 - J. Larson
A minor modification to the routine allflush_() in the mpeu module m_zeit.F90.
This modification was to appease the ftn compiler on the cray X-1.  The nature of
this change was to place parentheses around the array ztbf fed to the MPI function
MPI_gather().  This was to re-shape the array and enforce the alphabetical sorting
of the timing tag entries.