Changes from Bill Sacks at NCAR. Introduce new datatype AvSharedIndicesOneType and AvSharedIndices (a structure of two OneTypes). These types are used to hold the cross-indexed attribute indicies between two Avs. They are initialized with two new functions: SharedIndices and SharedIndicesOneType and cleaned with cleanSharedIndices and cleanSharedIndicesOneType.
The AvSharedIndices datatype can be used as an optional argument in the av:Copy routines so
that the index matching doesn't have to be done everytime the Copy routine is called.
This was shown to have substantial overall time savings in a case where Copy was frequently called.
The user must initialize one AvSharedIndices type for each Av pair.  Order matters.