Adding a systemPaths directory which contains the location of the NETCDF and PNETCDF installations for various compilers on YS. Also adding CMake infrastucture to automatically determine which system is being used and the compiler vendor. Then the correct "system file" is read to determine the correct path of dependencies. In this branch there is no need to use environment variables to set these locations.
The idea here is that certain machines will be "officially supported" since the locations of dependencies can get very complicated on supercomputers and it is probably too much for a user or for CMake to find the correct location. For other machines, such as Mac/Darwin, we can have CMake search standard locations for installations (like /usr/lib or /opt/local/). Otherwise, the user is free to specify the paths themselves on the configure (cmake) line. 




git-svn-id: file:///Users/mataylo/repos/homme/branches/cmake-2527@2535 b670e762-8b1b-0410-9e79-e6a6fc92c689

Former-commit-id: 0b91807a342d14cb8b5188aa987b8474e0d2889a