Updated the gridplot.ncl script get num_el_blk directly from netCDF file rather than using the dimsize of a variable that CUBIT uses but isn't really needed by HOMME (Paul U has a tool to hand-generate exodus files and he omits the varaible I had been using to determine the number of element blocks). Also, I realized this script had been using hard tabs instead of soft tabs... so I went ahead and changed that.
The change to refine_montain.jou is just to generate a grid where the coarse resolution is ne=9. Not actually a useful change, I just needed to test something with roughly the number of points found in the mountain_9_x2 mesh. As with mountain_3_x#, the mountain_9 case doesn't really refine around the mountain.



git-svn-id: file:///Users/mataylo/repos/homme/trunk@2822 b670e762-8b1b-0410-9e79-e6a6fc92c689

Former-commit-id: be97a60efc74af275b452558011ba3e5c65b1c33