moved 'ref2sphere' from coordinate_system_mod.F90 to cube_mod.F90
(ref2sphere is the routine that defines the mapping - it is not a 
coordinate transform, so I think it belongs in cube_mod.F90)

this minor code reorg is to allow us to introduce a new map that will
map element edges to to great circle arcs.  HOMME's existing equi-angular map
only maps coordinate lines to great circles.




git-svn-id: file:///Users/mataylo/repos/homme/trunk@2883 b670e762-8b1b-0410-9e79-e6a6fc92c689

Former-commit-id: 39d1ba8fea80886fc15b5e606fd20b18f8090627