Cleaning up reconstructed forcing fields.
Created a scratch var_struct in Registry.xml which can be used to house
scratch variables.

Moved normalVelocityForcingReconstruct fields into the scratch
var_struct, so they can be allocated/deallcated around a call to
mpas_reconstruct.

Created windStressZonal and windStressMeridional variables in the
diagnostics var_struct which are used to write (to output files) the
surface zonal and meridional wind stress.