Addition of Gent-McWilliams eddy parameterization.
These changes encompass work by Qingshan Chen, Todd Ringler, and Mark
Petersen.  The GM eddy parameterization is used to add the effects of
eddy variability at low resolution.  It includes the addition of a Bolus
velocity to the tracer transport, and a Redi diffusion of tracers that
occurs along isopycnals.