Fixing tracer loop in split explicit.
In split explicit, there was a loop intended to be over temperature and
salinity. The indicies for temperature and salinity were hard coded to
be 1 and 2, rather than using the start and end indices for the
array_group that temperature and salinity are part of.

This commit changes the explicit indicies to the start/end indices.

Fixes #4