Fixing an issue with IO Types for CESM use.
The PIO io_type definition within MPAS was defined within a subroutine.
This commit moves it to a module level variable, and provides a
set routine to set it, rather than having it be set through the init
call.

CESM determines what the io_type would be for a particlar run, so
instead of setting it internally we need to allow CESM to set it.