LI: Refactored vertical_remap
This commit refactors the vertical_remap subroutine to perform all
calculations truly column-wise, so it can be vectorized using
OpenMP over either blocks or cells.  This eliminates the needs for
scratch variables in this subroutine if OpenMP threading was to be
implemented in the future.
In my implementation I have avoided branching statements (if/where) and
instead used a small value 'eps' to avoid division by zero where there
is no ice.  This implementation then uses the cellMask to subsequently
zero out regions with no ice.  To facilitate this, I added new functions
and interface to the mask module that return 0/1 integers after querying
a mask rather than true/false.