change order and merging of cvmix parameterizations to match that done with POP.
bug in cvmix was not handling the use case of passing in cvmix_kpp_params.
so now we do not pass in this handle and, instead, allow cvmix to save the parameter settings.

cvmix kpp is returning plausible results for stable and unstable surface forcing.
remaining task is to implement nonlocal forcing function returned from kpp.