LI: Fix slope calculation in SIA
This commit fixes an inaccurate representation of ice sheet surface
slope in the SIA velocity solver.  Previously I had only been using the
slope in the normal direction but I have corrected that to also make use
of the total magnitude of the slope:
|grad s|^(n-1) * grad s  (see updated documentation for details)

I calculate the component of the slope tangent to an edge by first
averaging upperSurface from cells to vertices and then calculating the
slope between the two vertices adjacent to the edge.
The upperSurface value on vertices currently is only used for calculating
the tangent component of slope on edges for the SIA velocity solver.

Averaging from cells to vertices uses barycentric interpolation.
Note: this method will not work for obtuse triangles (where the
vertex is outside of the triangle formed by the neighboring cell
centers.)

This commit also corrects a bug in my previous implementation that was
using the gradient in the thickness instead of the surface slope!

This more accurate slope calculation brings the MPAS SIA results for the
EISMINT-1 mm test to be within the range of results of models in that
study (it previously had a dome center thickness that was too high).
This fix also results in less noisy SIA velocities, which make Greenland
simulations tractable.
For the Halfar dome after 200 years, the RMS error dropped from 14.3 m
to 7.5 m.