Adding clean as a dependency of parse.
This is because a core might choose to modify the NAMELIST_SUFFIX in a
manner that might not force a rebuilding of the parse executable. If the
parse executable is not rebuilt in this case, the namelist that is
generated by parse will still have the old suffix. So now, parse is
*always* rebuilt, since it's very minor in the over all compile time.