Updating registry to properly handle logicals.
Previously the test to see if a logical's default value was true or
false was incorrect if the default value was specified using Fortran
syntax (i.e. .false. and .true.), but was correct if they were specified
as true and false.

Registry has been updated to properly handle both cases.