POOLS: Adding pools to MPAS. Framework and cores.
This commit changes several portions of the MPAS code. It is a
combination of work done by Michael Duda, Matt Hoffman, and , Doug Jacobsen.

In the framework, the follow changes occured.
   - The pool data types were defined. Currently they are in the
     mpas_grid_types module.
   - Registry was updated to only generate code for pool data
     structures.
   - The framework was modified to generate and use pool data structures
     for: var_structs, packages, and namelist records / options.

Each core has additionally been updated to use pools. Results should be
bit-identical to the previous commit for every core.