Add option to use either USGS or MODIFIED_IGBP_MODIS_NOAH land use data in MPAS-Atmosphere.
When running the init_atmosphere core, users may select between two land use datasets by
setting the namelist option config_landuse_data to either 'USGS' or 'MODIFIED_IGBP_MODIS_NOAH'.
This choice is stored in the initial condition file in the string variable 'mminlu', from
which it is read by the atmosphere core. Users therefore do not need to specify the land use
dataset when running the atmosphere model.

For backwards compatibility with static and existing initial condition files, the land use dataset
is assumed to be USGS if the 'mminlu' variable is not found or contains no information.

Changes contributed by Dominikus Heinzeller (IMK).