Add an MPAS_Clock_type to the domain type.
The stream manager module will require access to the simulation clock when it
is initialized, and therefore, having the simulation clock as a module variable
in the mpas_core module is not feasible. So, we convert the module variable that
was previously an MPAS_Clock_type to a pointer to an MPAS_Clock_type, and set this
pointer to point to domain % clock at the beginning of the mpas_core_init() routine.