Adding mpas_stream_list module.
This module provides two new types that represent a linked list of
streams.

The streams can be added, removed, and searched within the linked list.
This is to allow a stream manager to keep a list of streams that are
active, so it doesn't need to re-create streams that are already active.