Obtain filename used by bootstrapping from the XML stream file and namelist options.
Rather than constructing the filename that will contain mesh information using
the namelist options config_do_restart, config_start_time, config_input_name, and
config_restart_name, the filename that will provide mesh information to the MPAS
bootstrapping routine is now obtained from the cores, acting upon information in
the namelist, and the XML stream parsing code. The key changes include:

1) Add new routine mpas_core_get_mesh_stream() in every core to return the
stream name that will contain mesh information.

2) Add new routine xml_stream_get_filename() in framework that, given the name
of a stream, will return the filename template associated with that stream.

3) Pass a mesh filename to mpas_bootstrap_framework(), rather than obtaining
the filename as in the past.

4) Have the subdriver query the mesh stream name from the core, convert the
stream name to a filename template, convert the filename template to an actual
filename, and call the bootstrapping routine.