Fix broken surface updates in MPAS-Atmosphere.
Before the implementation of pools, the subroutine mpas_io_input_init() was
responsible for both creating a stream and adding the fields that belonged to
the stream. With the implementation of pools, the code to add fields to a stream
was moved to a separate subroutine, mpas_add_input_fields(), and the new
subroutine call was not added for the surface update stream in MPAS-Atmosphere.

Now, in mpas_atm_mpas_core.F, we add the call

          call mpas_add_input_fields(domain, sfc_update_obj)

immediately after the call to

          call mpas_io_input_init(sfc_update_obj, domain % dminfo)

.