Change same grid mapping tolereance
This change is needed to run the RRM with the arm_x8v3 grid.
Otherwise, this grid will fail one the mapping tolereance checks.

This change should not be pushed to to ACME master until we understand this better

*********1*********2*********3*********4*********5
Longer commit message body describing the commit.
Can contain lists as follows:
	* Item 1
	* Item 2
	* Item 3

A good commit message should be written like an email, a subject followed by a
blank line, followed by a more descriptive body.

Can also contain a tag at the bottom describing if the commit is BFB, Non-BFB,
or Climate changing:
[BFB]
[Non-BFB]
[CC]