Implement a capability for updating field pointers in an already constructed I/O stream.
In the stream manager, we may need to read a field from file into any specified time level
of that field. One possible approach would involve destroying the existing stream and rebuilding
a new stream that points to the new specified time level for fields, while the more efficient
approach would involve simply updating which field types are pointed to in a stream before fields
are read into the fields.

This field update capability is provided by a new routine MPAS_streamUpdateField in the
mpas_io_streams module.