Eliminating 'failure' messages when calling 'svn info' since this is not a subversion repository
'git describe' and 'svn info' are both called in the logic that archives provenance
for the code with performance data, to provide something that will work both with
ACME and with non-ACME versions of CESM. This generates 'failed' messages depending on
whether this is a git or a subversion repository.

Logic is added to test whether .git or .svn exist at the repository root before calling
git or svn commands, respectively. (This appears to be a standard approach.)