Delaying instantiation of CESM variables in test scripts (part 2)
When the definition of CESMSCRATCHROOT includes CESM variables,
in particular $PROJECT, in cesm1_3_beta10 these are expanded
immediately, before the CESM variables are defined. A previous
update fixed this for the scripts create_newcase and create_test,
but not create_clone. Bill Sacks redesigned the previous
implementation and also added support in create_clone (as
part of the ACES4BGC project). This commit ports this work
into the ACME code base.