Add initial implementations of MPAS_stream_mgr_write() and MPAS_stream_mgr_read() to the mpas_stream_manager module.
These two routines are believed to be working with all of their optional
arguments except the following:
* MPAS_stream_mgr_write(): 'mgLevel' and 'forceWriteNow'
* MPAS_stream_mgr_read(): 'mgLevel' and 'rightNow'; the 'when' argument may be
  working by chance but has not been tested

When one alarm is associated with two or more streams, and we are only reading
or writing a subset of those streams, we can't just reset the alarm, since that
would cause us to not handle the other streams in a subsequent call to the read
or write routines. To work around this issue, a new routine,
MPAS_stream_mgr_reset_alarms(), has been added to the stream manager; after all
streams have been handled this routine should be called by user code to reset
the alarms that were ringing.

Additional, private routines that have been added to the mpas_stream_manager_module
include read_stream(), write_stream(), and update_stream().