Bug fix: don't exit stream loop in MPAS_stream_mgr_{read,write} if a stream of the wrong direction is found.
This bug only affected calls to these two routines in which the optional streamID argument was not present,
in which case we loop over all streams in the manager.

The previous, incorrect logic in the MPAS_stream_mgr_read() routine would exit when an output-only stream
was found; similarly, the logic in MPAS_stream_mgr_write() would exit upon encountering an input-only
stream. Now, we simply skip over streams of the wrong direction.