Pass an MPAS_streamManager_type to mpas_core_init(), mpas_core_run(), and mpas_core_finalize().
Currently, an MPAS_streamManager_type is defined as a module variable in the mpas_subdriver module;
this instance is set up in the mpas_init() routine after bootstrapping the framework, but was not
previously available to MPAS cores. Now, we pass this manager instance to all cores as an argument
to the mpas_core_init(), mpas_core_run(), and mpas_core_finalize() routines.

After all cores have migrated to using the stream manager for reading and writing streams, including
initial conditions, we should be able to remove routines like mpas_input_state_for_domain() and
mpas_output_state_for_domain() from the mpas_io_input and mpas_io_output modules.