Updating the ocean core for runtime streams
This commit updates the forward model and the analysis model of the
ocean core to work with multiple streams.

It changes the structure of analysis members, and adds logic that is
based on the stream manager. Analysis members will only be computed if
the stream manager says they are to be written on a given time step.