Enable MPAS to be built with PIO libraries that enabled only one of NetCDF or Parallel-NetCDF.
Until now, it was assumed that PIO was built with support for both NetCDF and Parallel-NetCDF,
and the settings of CPPINCLUDES, FCINCLUDES, and LIBS in the top-level Makefile reflected this
fact.

Now, appropriate paths and libraries are only added to CPPINCLUDES, FCINCLUDES, and LIBS according
to the presence of environment variables NETCDF and PNETCDF.

Note: There is still work to be done inside the I/O layers of MPAS to allow users to select either
the NetCDF or Parallel-NetCDF libraries for reading and writing; the changes in this commit only
address issues in the Makefile.