Handle streams with fields tied to inactive packages
Registry is updated to create all fields as far as the allFields pool is
concerned. Each field has isActive set to false, unless it is activated,
in which case isActive is set to true and the field is also added to
it's respective var_struct.

If the field indeed is not activated, it is not in it's respective
var_struct, but it is still in allFields.

Streams now check isActive before adding the field to stream. This
allows typos in field names to be caught, and deactivated fields to be
skipped without throwing an error