When reading from a stream with maxRecords=0, use MPAS_seekStream() to find proper record to read from.
Previously, the logic in read_stream(), which is called from MPAS_stream_mgr_read(),
assumed that, if all records for a stream were packed in the same file, then we should
simply begin reading from the first index. This logic gives incorrect results, e.g.,
for forcing streams, when a run begins at a time later than the first time in the stream.

Note: A similar change might be applicable in write_stream() for situations where we
are writing to an existing file.