Silence harmless pool messages in MPAS_stream_mgr_add_field() for debug builds.
In the MPAS_stream_mgr_add_field() routine, we expect that, unless a field was already added
to a stream, the call to check whether that field is in the field pool for a stream will fail.
For debug builds, however, the pool query routine generates a large number of misleading error
messages.

Now, before checking that a field isn't already part of a stream, we first query the current pool
error reporting level, set the level to MPAS_POOL_SILENT, perform the query, then reset the level
to its original value.