Removing records_per_file with filename_interval in streams
filename_interval defines the interval between filename changes. It is
used to create new files when writing a stream, or to determine which
file is likely to contain a record when reading a stream.

filename_interval is optional, and can take the values of a timestamp
(i.e. "YYYY-MM-DD_hh:mm:ss"), "none", "output_interval", or
"input_interval".

If it is not specified, the following rules are used to determine the
value of filename_interval.
 - If The stream type contains input (i.e. input or input;output) and
   input_interval is an actual interval (i.e. not initial_only or none)
    - filename_interval defaults to input_interval
 - If The stream type contains output (i.e. output or input;output) and
   output_interval is an actual interval (i.e. not initial_only or none)
   and input interval is not already used as the default
    - filename_interval defaults to output_interval
 - Otherwise, filename_interval defaults to none

 filename_interval is not updated over the course of a simulation when
 output_interval or input_interval are updated, even if the value of
 filename_interval is "output_interval" or "input_interval".