Automatically handle writing to a pre-existing file in the low-level MPAS I/O module (mpas_io).
Before calling PIO_createfile, we first inquire whether the file to be created exists; if so,
we instead call PIO_openfile. Besides this simple change, we also need to avoid trying to redefine
variables that are already in the file.

With this change, we can now have a stream append to an existing file.