Add optional StartTimeIn to mpas_get_timeInterval
For time intervals that require months and/or years, ESMF needs to know the start
time to get a time interval in any format besides the string format.  To allow
getting time intervals for such cases, this commit adds an optional argument to
mpas_get_timeInterval called StartTimeIn which is the start time of the interval.

If present, this argument is passed through to ESMF_TimeIntervalGet.