LI: Use timeString for dt
With the newer version of ESMF, intervals that include months and years are now
supported.  Therefore, this commit converts the dt specified in the namelist
from the current implementation of either in units of years or seconds to now
only support a time string of format 'YYYY-MM-DD_HH:MM:SS'.  Within
mpas_li_mpas_core.F the time step is represented as MPAS_TimeInterval_type. This
type is converted to a real type in units of seconds before being passed to the
time stepper.

Note that items in the format string may be dropped from the left if not needed,
and that values for the items may be larger than what is needed to form the next
unit (e.g., seconds may be >=60).

I have removed li_deltat and li_secondsInYear from the code entirely since these
are no longer needed with this change.