Adding support for Depends.$mach.$compiler files to fix support for cray, pgi, pgicuda, and cray on Titan
An earlier check-in adding support for the Cray and Intel compilers
on Titan used a naming convention for the Depends file that is not recognized.
In fact there is no support for Depends files that are dependent on both
machine and compiler. This capability is being added with the naming convention
Depends.$mach.$compiler. When a Depends file with both $mach and $compiler exist,
Depends.$mach and Depends.$compiler are ignored.

The Depends files from the previous check-in are also renamed to reflect the new design.