Restore restart_timestamp file, add "forcing"
The conversion to streams deleted the functionality to use a restart_timestamp
file to indicate the time at which to begin a restart.  These changes restore
that functionality.  Patterned after commit 5c86f6 in branch ocean/v3.0_cleanup.

This commit also adds an input-read near the end of each time step to allow
time-varying inputs (i.e., forcing).  Currently this is only piecewise-constant.