Change default pe layout for I case on blues
The ICLM45BGC ne30_g16 case on blues requires 2 nodes to run or it runs
out of memory.  Update config_pes.xml so that this case on blues
asks for 32 tasks (2 nodes).   Had to move a default setting
for I cases before the blues setting because last match wins
in config_pes.xml.

[BFB]