Adding a script to maintain acme_developer suite.
This script eases the burden inherent in maintaining a set of "floating"
tests on a large number of machines with individualized compilers, a need that
is not met by the hierarchical structure of the CESM testing system.

*********1*********2*********3*********4*********5*********6*********7**
Longer commit message body describing the commit.
Can contain lists as follows:
	* Item 1
	* Item 2
	* Item 3

A good commit message should be written like an email, a subject
followed by a blank line, followed by a more descriptive body.

Can also contain a tag at the bottom describing if the commit is BFB,
Non-BFB, or Climate changing:
[BFB]
[Non-BFB]
[CC]