New PRNG, bug fixes, namelist working
This introduces some bug fixes in the code from the last commit (syntax stuff). Also, we've put in a new RNG thanks to Pat Worley requesting it from another project and adapting it to this code. There is a severe bug in PGI's PRNG, in which its "random" numbers are repeated quite often, and they are orders of magnitude too small, and their order of magnitude depends on the seed value. So the new PRNG (which is not the fortran intrinsic) fixes this problem, and its repeatable between different compilers, which is nice. We've included a flag (new_random) in the cam_control namelist, which turns the new PRNG on or off. The seed_custom and seed_clock inputs are the same as before. I've added the new namelist parameters to namelist_definition.xml so they can be specified in user_nl_cam.