Fixing issues with when using multiple blocks
Prior to a diagnostic_solve call, forcingPool was not associated with
the correct pool.

Two arrays (normalVelocityNew and maxLevelEdgeTop) were previously used
in a block loop without retrieving pointers to them from the current
block's pool.

When run with multiple blocks, this caused an array out of bounds issue
becuase the arrays would be associated to the last block's arrays. Which
might have a different size than the other blocks in the blocklist.