wait_for_tests: Add signal handling
When Jenkins times out, we want to see the most recent test results
in the job output. This commit changes wait_for_tests to not simply
die when it gets hit with a SIGINT or SIGTERM. Instead, the program
will print all the test statuses and then exit.

In order to ensure that test results are captured, the output of
wait_for_tests should be redirected to a file.