Add a function for subcell fluxes of the divergence.
This allows the computation of the integral of the divergence
of a vector field to be replaced by the boundary integral.
The unit cells are the mesh elements chopped up into a regular
checkerboard of subcells. Input is the field on the element.
What is returned is an N by N by 4 array giving the fluxes
for each sub-element. Note that this is redundant since
neighboring elements have the same fluxes on shared sides.
It is a thing Mark wanted.

Also added a unit test that can be enabled by changing
an #if 0 in sweq_mod.F90.




git-svn-id: file:///Users/mataylo/repos/homme/trunk@4409 b670e762-8b1b-0410-9e79-e6a6fc92c689

Former-commit-id: 198e1a4eeceb67b8f32ff5954760153e3855cba7