Updating the bootstrapping process to handle stream attributes properly
Previously when framework was bootstrapping itself based on the mesh
stream, it would determine the filename to read from only based on the
start time. However, when a restart run was occuring, the start time of
the simulation might not define a valid filename to read from.

Now framework determines the filename to read the fields required to
bootstreap from by examining the attributes of the mesh stream:
    - filename_template
    - reference_time
    - filename_interval

This allows a restart run to look in a file before start_time to find
the time start time when bootstrapping.

Also writing the name of the file bootstrapping is performed with.