Fixing directory creation when nested in non-writable directories
Previously, directory creation within streams would fail if any of the
directories was not writable. In reality, only the directories we want
to create directories and files in need to be writable. This commit
changes directory creation to only fail if a directory we need to be
able to write in is not writable.