Always compute the difference between two time instants in terms of days and seconds.
Presumably for efficiency reasons, the code in ESMF_TimeMod.F90 that computes
a time interval as the difference between two time instants would express the interval
in terms of an integral number of years and months if the days and seconds of
the two time instants were the same. However, the time interval comparison routine
timeintcmp() cannot compare two time intervals if the years and seconds don't match,
which causes problems. For example, the following would not work in MPAS:

      call mpas_set_time(curr_time=startTime, dateTimeString='2014-12-01_00:00:00')
      call mpas_set_time(curr_time=stopTime, dateTimeString='2016-06-01_00:00:00')

      call mpas_set_timeInterval(compInterval, dt=(86400.0_RKIND * 546.0))
      runDuration = stopTime - startTime
      if (runDuration > compInterval) then
         ! do something
      else
         ! do something else
      end if

To avoid this issue, we now always compute the time interval in terms of days
and seconds.