Adding an error for reads of non-existent files
When a file is read in MPAS, but the file doesn't exist, the model gave
a generic PIO error that doesn't inform the user that the file they told
the model to read doesn't exist.

This commit adds a new error code and message that describe that the
file doesn't exist so a user can fix the issue.