LI: Add option to do vector reconst. w/ extern. dycores
Recently the external dycore interface was modified so that they pass
back u,v velocities on cell centers (which are the native locations of
their FEM solution).  This commit adds an option that lets MPAS
overwrite those values with ones calculated by MPAS using framework's vector
reconstruction routines based on the values of normalVelocity supplied
by the external dycore.  This provides a way to test the calculation of
normalVelocity in the interface.